ml_model/requirements.txt
tensorflow
tensorflow-macos
pillow
scipy 